Reusing <jena-base>/jena.cmd classes in <jena-core>
